Move LoadingModal
