+++ src/components/LoadingModal.js
@@ -1,7 +1,7 @@
 import React from 'react'
-import { CoolAlert } from '../CoolAlert'
+import { CoolAlert } from './CoolAlert'
 
-import translate from '../../../locales/i18n'
+import translate from '../../locales/i18n'
 
 const LoadingModal = ({ show }) => {
     return (
