+++ src/screens/inicio/Register/index.js
@@ -35,7 +35,7 @@ import {
 import { PageTitle } from './styles'
 
 import InstitutionSelector from '../../../components/Groups/InstitutionSelector'
-import LoadingModal from '../../../components/Groups/LoadingModal'
+import LoadingModal from '../../../components/LoadingModal'
 import translate from '../../../../locales/i18n'
 import { scale } from '../../../utils/scalling'
 import { UserIcon } from '../../../img/imageConst'
