+++ src/screens/principal/EditarPerfil/index.js
@@ -33,7 +33,7 @@ import {
 import { Delete } from './styles'
 
 import InstitutionSelector from '../../../components/Groups/InstitutionSelector'
-import LoadingModal from '../../../components/Groups/LoadingModal'
+import LoadingModal from '../../../components/LoadingModal'
 import translate from '../../../../locales/i18n'
 import { scale } from '../../../utils/scalling'
 import {
