+++ src/screens/principal/NovoPerfil/index.js
@@ -30,7 +30,7 @@ import {
 } from '../../../components/NormalForms'
 
 import InstitutionSelector from '../../../components/Groups/InstitutionSelector'
-import LoadingModal from '../../../components/Groups/LoadingModal'
+import LoadingModal from '../../../components/LoadingModal'
 import translate from '../../../../locales/i18n'
 import { scale } from '../../../utils/scalling'
 import {
