+++ src/screens/principal/Vigilancia/index.js
@@ -23,7 +23,7 @@ import {
 } from '../../../components/NormalForms'
 import { Title, BodyText } from './styles'
 
-import LoadingModal from '../../../components/Groups/LoadingModal'
+import LoadingModal from '../../../components/LoadingModal'
 import translate from '../../../../locales/i18n'
 import { useUser } from '../../../hooks/user'
 import { scale } from '../../../utils/scalling'
